Update UI.py
The updated UI uses customtkinter instead of tkinter for a more modern UI feel. Additionally, I used Pillow to help display the image of the movie posters from the database to the UI to help with searching for movies. You will need to download/install customtkinter and Pillow for the UI to be functional, if it is not already installed on your device